Batch - Move tests into separate files
